# api/openapi.yaml
openapi: 3.0.3
info:
  title: Spillway admin API
  version: 0.4.0
  description: |
    The API behind the dashboard. Its only caller is the dashboard's server-side layer, which signs users in
    with POST /admin/login and forwards the returned token as `Authorization: Bearer <token>`. The browser
    never sees the token. TypeScript types for web/ are generated from this file, and a Go test checks that
    the real responses match it.

    Each operation says which roles may call it with `x-access`: `public`, `viewer` (any signed-in user)
    or `admin`. The Go router registers the same access level, and a test fails if they disagree.
paths:
...
  /admin/runs:
    get:
      operationId: runsList
      summary: Runs, newest first, each with a strip of its latest steps
      x-access: viewer
      security: [{ bearer: [] }]
      parameters:
        - { name: state, in: query, required: false, description: "active is every run not finished; finished is runs that ended inside `hours`", schema: { type: string, enum: [active, finished] } }
        - { name: status, in: query, required: false, schema: { type: string, enum: [queued, running, waiting_tool, waiting_human, sleeping, succeeded, failed, cancelled] } }
        - { name: hours, in: query, required: false, schema: { type: integer, enum: [1, 24, 168], default: 24 } }
        - { name: q, in: query, required: false, description: Only runs whose task contains this text (ignoring case), schema: { type: string, maxLength: 100 } }
        - { name: key_id, in: query, required: false, description: Only runs of this API key, schema: { type: string, format: uuid } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: cursor, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: One page
          content:
            application/json:
              schema:
                type: object
                required: [runs, next_cursor]
                properties:
                  runs:
                    type: array
                    items: { $ref: "#/components/schemas/RunItem" }
                  next_cursor: { type: string, nullable: true }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: runsCreate
      summary: Start a run from the dashboard
      description: The run is made under the built-in playground key and charged to it. Tools are not available from the dashboard yet.
      x-access: admin
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RunStart" }
      responses:
        "202":
          description: Queued
          content:
            application/json:
              schema:
                type: object
                required: [id, status]
                properties:
                  id: { type: string, format: uuid }
                  status: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/runs/{id}:
    get:
      operationId: runGet
      summary: One run, in the shape of a list row
      x-access: viewer
      security: [{ bearer: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: The run
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RunItem" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
...
  responses:
...
    BadRequest:
      description: The request is not valid
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: No token, a bad token, or an expired one
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: The signed-in user is a viewer and this needs an admin
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such resource
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
...
    StripNode:
      type: object
      required: [kind, state]
      properties:
        kind: { type: string, enum: [model, tool] }
        state: { type: string, enum: [done, current, failed, waiting] }
    Strip:
      type: object
      required: [steps, more]
      properties:
        steps: { type: array, items: { $ref: "#/components/schemas/StripNode" }, description: "At most 16: the latest ones" }
        more: { type: integer, description: How many earlier steps are not shown }
    RunItem:
      type: object
      required: [id, status, goal, key, model, step_count, cost_usd, created_at, finished_at, duration_ms, wake_at, failure_reason, strip]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [queued, running, waiting_tool, waiting_human, sleeping, succeeded, failed, cancelled] }
        goal: { type: string }
        key: { type: string }
        model: { type: string }
        step_count: { type: integer }
        cost_usd: { type: string }
        created_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time, nullable: true }
        duration_ms: { type: integer, format: int64 }
        wake_at: { type: string, format: date-time, nullable: true }
        failure_reason: { type: string, nullable: true }
        strip: { $ref: "#/components/schemas/Strip" }
    RunStart:
      type: object
      required: [input]
      additionalProperties: false
      properties:
        input: { type: string }
        system: { type: string }
        model: { type: string, description: A policy name or model id; default is the default policy }
        limits:
          type: object
          additionalProperties: false
          properties:
            max_steps: { type: integer, minimum: 1 }
            max_cost_usd: { type: number }
            deadline_seconds: { type: integer, minimum: 1 }
        metadata: { type: object }
...